Corrigindo detalhes de UX

+++ Projetos - Programação/Pizzaria programa/Pizzaria.Backend/classes/Produtos.cs
@@ -5,7 +5,7 @@ namespace Pizzaria.Backend.classes {
         public int Id;
         public string Nome;
         public string Descricao;
-        public float Preco;
+        public double Preco;
         public string Categoria;
         public DateTime DataCriacao;
     }
